# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/markets/top-movers": {
      "get": {
        "summary": "Get Top Movers Endpoint",
        "description": "Get markets with highest absolute price delta (top movers).\nReturns markets sorted by absolute price change since last refresh.",
        "operationId": "get_top_movers_endpoint_markets_top_movers_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 30,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TopMoversResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/search-slug": {
      "get": {
        "summary": "Get Market By Slug Endpoint",
        "description": "Get a market by its slug from Polymarket Gamma API.\n\nOfficial docs: https://docs.polymarket.com/api-reference/markets/get-market-by-slug",
        "operationId": "get_market_by_slug_endpoint_markets_search_slug_get",
        "parameters": [
          {
            "name": "slug",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Market"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/trades": {
      "get": {
        "summary": "Get Market Trades Endpoint",
        "operationId": "get_market_trades_endpoint_markets_trades_get",
        "parameters": [
          {
            "name": "condition_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Condition Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Trade"
                  },
                  "title": "Response Get Market Trades Endpoint Markets Trades Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/search": {
      "get": {
        "summary": "Get Search Markets Endpoint",
        "description": "Search for markets using Polymarket Gamma API.\nReturns events with their markets, filtered to only active markets.",
        "operationId": "get_search_markets_endpoint_markets_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/search": {
      "get": {
        "summary": "Get Search Profiles Endpoint",
        "description": "Search for profiles using Polymarket Gamma API.",
        "operationId": "get_search_profiles_endpoint_profiles_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/public-profile": {
      "get": {
        "summary": "Get Public Profile Endpoint",
        "description": "Get a public profile by address from Polymarket Gamma API.",
        "operationId": "get_public_profile_endpoint_profiles_public_profile_get",
        "parameters": [
          {
            "name": "address",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Address"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublicProfile"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/by-condition-ids": {
      "get": {
        "summary": "Get Markets By Condition Ids Endpoint",
        "description": "Get multiple markets by their condition IDs from Polymarket Gamma API.\n\nOfficial docs: https://docs.polymarket.com/api-reference/markets/list-markets",
        "operationId": "get_markets_by_condition_ids_endpoint_markets_by_condition_ids_get",
        "parameters": [
          {
            "name": "condition_ids",
            "in": "query",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "description": "Repeat the query param per ID",
              "title": "Condition Ids"
            },
            "description": "Repeat the query param per ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Market"
                  },
                  "title": "Response Get Markets By Condition Ids Endpoint Markets By Condition Ids Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Markets By Condition Ids Endpoint",
        "description": "POST variant that accepts condition IDs in the JSON body.",
        "operationId": "post_markets_by_condition_ids_endpoint_markets_by_condition_ids_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_markets_by_condition_ids_endpoint_markets_by_condition_ids_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Market"
                  },
                  "title": "Response Post Markets By Condition Ids Endpoint Markets By Condition Ids Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/top-holders-pnl": {
      "post": {
        "summary": "Get Top Holders Pnl Endpoint",
        "description": "Get PnL data for holders.\n\nAccepts a list of TopHolder objects and returns them with PnL data (avgPrice, realizedPnl, totalBought).",
        "operationId": "get_top_holders_pnl_endpoint_markets_top_holders_pnl_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopHoldersPnlRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopHolderPnl"
                  },
                  "type": "array",
                  "title": "Response Get Top Holders Pnl Endpoint Markets Top Holders Pnl Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/top-holders-wallet-info": {
      "post": {
        "summary": "Get Top Holders Wallet Info Endpoint",
        "description": "Get wallet information for holders.\n\nAccepts a list of TopHolder objects and returns them with wallet info (walletCreatedAt, walletLastTransfer, walletBalance).",
        "operationId": "get_top_holders_wallet_info_endpoint_markets_top_holders_wallet_info_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TopHoldersWalletInfoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TopHolderWalletInfo"
                  },
                  "type": "array",
                  "title": "Response Get Top Holders Wallet Info Endpoint Markets Top Holders Wallet Info Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "summary": "Get Event By Id Endpoint",
        "description": "Get an event by its ID from Polymarket Gamma API.\n\nOfficial docs: https://docs.polymarket.com/api-reference/events/get-event-by-id",
        "operationId": "get_event_by_id_endpoint_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/markets/clarifications": {
      "get": {
        "summary": "Get Clarifications Endpoint",
        "description": "Get all updates for a questionID and owner from the UMA CTF Adapter contract on-chain.\n\nCalls getUpdates(bytes32 questionID, address owner) on contract 0x6A9D222616C90FcA5754cd1333cFD9b7fb6a4F74.\nResults are cached in Redis for 24 hours.",
        "operationId": "get_clarifications_endpoint_markets_clarifications_get",
        "parameters": [
          {
            "name": "question_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Question ID as hex string (bytes32)",
              "title": "Question Id"
            },
            "description": "Question ID as hex string (bytes32)"
          },
          {
            "name": "owner",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Owner address",
              "title": "Owner"
            },
            "description": "Owner address"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AncillaryDataUpdate"
                  },
                  "title": "Response Get Clarifications Endpoint Markets Clarifications Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_address}/pnl-with-markers": {
      "get": {
        "summary": "Get User Pnl With Markers",
        "description": "Returns user PnL points and chart-aligned markers (swings and trade clusters).",
        "operationId": "get_user_pnl_with_markers_users__user_address__pnl_with_markers_get",
        "parameters": [
          {
            "name": "user_address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Address"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "1m",
              "title": "Interval"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PnlWithMarkersResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Market": {
        "properties": {
          "conditionId": {
            "type": "string",
            "title": "Conditionid"
          },
          "question": {
            "type": "string",
            "title": "Question"
          },
          "questionId": {
            "type": "string",
            "title": "Questionid"
          },
          "marketMakerAddress": {
            "type": "string",
            "title": "Marketmakeraddress"
          },
          "submitted_by": {
            "type": "string",
            "title": "Submitted By"
          },
          "icon": {
            "type": "string",
            "title": "Icon"
          },
          "image": {
            "type": "string",
            "title": "Image"
          },
          "outcomes": {
            "type": "string",
            "title": "Outcomes"
          },
          "outcomePrices": {
            "type": "string",
            "title": "Outcomeprices"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "token1": {
            "type": "string",
            "title": "Token1"
          },
          "token2": {
            "type": "string",
            "title": "Token2"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "liquidity": {
            "type": "number",
            "minimum": 0.0,
            "title": "Liquidity"
          },
          "volume": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume"
          },
          "volume24hr": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume24Hr"
          },
          "volume1wk": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume1Wk"
          },
          "volume1mo": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume1Mo"
          },
          "volume1yr": {
            "type": "number",
            "minimum": 0.0,
            "title": "Volume1Yr"
          },
          "negRisk": {
            "type": "boolean",
            "title": "Negrisk"
          },
          "bestBid": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Bestbid"
          },
          "bestAsk": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Bestask"
          },
          "endDate": {
            "type": "string",
            "title": "Enddate"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "closed": {
            "type": "boolean",
            "title": "Closed"
          },
          "groupItemTitle": {
            "type": "string",
            "title": "Groupitemtitle"
          },
          "resolutionSource": {
            "type": "string",
            "title": "Resolutionsource"
          },
          "events": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MarketEvent"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Events"
          },
          "umaReward": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Umareward"
          },
          "clobRewards": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ClobReward"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clobrewards"
          },
          "rewardsMinSize": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewardsminsize"
          },
          "rewardsMaxSpread": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rewardsmaxspread"
          },
          "holdingRewardsEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holdingrewardsenabled"
          },
          "feesEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feesenabled"
          }
        },
        "type": "object",
        "required": [
          "conditionId",
          "question",
          "questionId",
          "marketMakerAddress",
          "submitted_by",
          "icon",
          "image",
          "outcomes",
          "outcomePrices",
          "slug",
          "token1",
          "token2",
          "description",
          "liquidity",
          "volume",
          "volume24hr",
          "volume1wk",
          "volume1mo",
          "volume1yr",
          "negRisk",
          "bestBid",
          "bestAsk",
          "endDate",
          "active",
          "closed",
          "groupItemTitle",
          "resolutionSource"
        ],
        "title": "Market"
      },
...
      "TopMoversResponse": {
        "properties": {
          "movers": {
            "items": {
              "$ref": "#/components/schemas/TopMover"
            },
            "type": "array",
            "title": "Movers"
          },
          "fetched_at": {
            "type": "string",
            "title": "Fetched At"
          }
        },
        "type": "object",
        "required": [
          "movers",
          "fetched_at"
        ],
        "title": "TopMoversResponse",
        "description": "Response for top movers endpoint."
      },
...
    }
  }
}